+++ test/Microsoft.AspNetCore.Server.Kestrel.Core.Tests/FrameRequestStreamTests.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.IO;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
 using Moq;
@@ -186,6 +185,302 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
             var stream = new FrameRequestStream();
             stream.StartAcceptingReads(null);
             Assert.Throws<ArgumentException>(() => { stream.CopyToAsync(Mock.Of<Stream>(), 0); });
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public async Task CanReadFromContentLength(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders { HeaderContentLength = "5" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+
+                var count = stream.Read(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                await readerTask;
+
+                count = stream.Read(buffer, 0, buffer.Length);
+                Assert.Equal(0, count);
+            }
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public async Task CanReadAsyncFromContentLength(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders { HeaderContentLength = "5" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+
+                var count = await stream.ReadAsync(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                await readerTask;
+
+                count = await stream.ReadAsync(buffer, 0, buffer.Length);
+                Assert.Equal(0, count);
+            }
+        }
+
+        [Fact]
+        public async Task CanReadFromChunkedEncoding()
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(HttpVersion.Http11, new FrameRequestHeaders { HeaderTransferEncoding = "chunked" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("5\r\nHello\r\n");
+
+                var buffer = new byte[1024];
+
+                var count = stream.Read(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                input.Add("0\r\n\r\n");
+
+                await readerTask;
+
+                count = stream.Read(buffer, 0, buffer.Length);
+                Assert.Equal(0, count);
+            }
+        }
+
+        [Fact]
+        public async Task CanReadAsyncFromChunkedEncoding()
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(HttpVersion.Http11, new FrameRequestHeaders { HeaderTransferEncoding = "chunked" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("5\r\nHello\r\n");
+
+                var buffer = new byte[1024];
+
+                var count = await stream.ReadAsync(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                input.Add("0\r\n\r\n");
+
+                await readerTask;
+
+                count = await stream.ReadAsync(buffer, 0, buffer.Length);
+                Assert.Equal(0, count);
+            }
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public void CanReadFromRemainingData(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders { HeaderConnection = "upgrade" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+
+                var count = stream.Read(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                Assert.False(readerTask.IsCompleted);
+            }
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public async Task CanReadAsyncFromRemainingData(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders { HeaderConnection = "upgrade" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask  = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+
+                var count = await stream.ReadAsync(buffer, 0, buffer.Length);
+                Assert.Equal(5, count);
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, count));
+
+                Assert.False(readerTask.IsCompleted);
+            }
+        }
+
+        [Theory]
+        [InlineData("keep-alive, upgrade")]
+        [InlineData("Keep-Alive, Upgrade")]
+        [InlineData("upgrade, keep-alive")]
+        [InlineData("Upgrade, Keep-Alive")]
+        public void CanReadFromConnectionUpgradeKeepAlive(string headerConnection)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(HttpVersion.Http11, new FrameRequestHeaders { HeaderConnection = headerConnection }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var readerTask = reader.StartAsync(body);
+
+                var stream = new FrameRequestStream();
+                stream.StartAcceptingReads(reader);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+                Assert.Equal(5, stream.Read(buffer, 0, buffer.Length));
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, 5));
+            }
+        }
+
+        [Theory]
+        [InlineData("keep-alive, upgrade")]
+        [InlineData("Keep-Alive, Upgrade")]
+        [InlineData("upgrade, keep-alive")]
+        [InlineData("Upgrade, Keep-Alive")]
+        public async Task CanReadAsyncFromConnectionUpgradeKeepAlive(string headerConnection)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(HttpVersion.Http11, new FrameRequestHeaders { HeaderConnection = headerConnection }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var readerTask = reader.StartAsync(body);
+
+                var stream = new FrameRequestStream();
+                stream.StartAcceptingReads(reader);
+
+                input.Add("Hello");
+
+                var buffer = new byte[1024];
+                Assert.Equal(5, await stream.ReadAsync(buffer, 0, buffer.Length));
+                AssertExtensions.Ascii("Hello", new ArraySegment<byte>(buffer, 0, 5));
+            }
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public async Task ReadFromNoContentLengthReturnsZero(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders(), input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                await readerTask;
+
+                var buffer = new byte[1024];
+                Assert.Equal(0, stream.Read(buffer, 0, buffer.Length));
+            }
+        }
+
+        [Theory]
+        [InlineData(HttpVersion.Http10)]
+        [InlineData(HttpVersion.Http11)]
+        public async Task ReadAsyncFromNoContentLengthReturnsZero(HttpVersion httpVersion)
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(httpVersion, new FrameRequestHeaders(), input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                input.Add("Hello");
+
+                await readerTask;
+
+                var buffer = new byte[1024];
+                Assert.Equal(0, await stream.ReadAsync(buffer, 0, buffer.Length));
+            }
+        }
+
+        [Fact]
+        public async Task CanHandleLargeBlocks()
+        {
+            using (var input = new TestInput())
+            {
+                var body = MessageBody.For(HttpVersion.Http10, new FrameRequestHeaders { HeaderContentLength = "8197" }, input.FrameContext);
+                var reader = new RequestBodyReader(input.PipeFactory.Create());
+                var stream = new FrameRequestStream();
+
+                stream.StartAcceptingReads(reader);
+                var readerTask = reader.StartAsync(body);
+
+                // Input needs to be greater than 4032 bytes to allocate a block not backed by a slab.
+                var largeInput = new string('a', 8192);
+
+                input.Add(largeInput);
+                // Add a smaller block to the end so that SocketInput attempts to return the large
+                // block to the memory pool.
+                input.Add("Hello");
+
+                var ms = new MemoryStream();
+
+                await stream.CopyToAsync(ms);
+                var requestArray = ms.ToArray();
+                Assert.Equal(8197, requestArray.Length);
+                AssertExtensions.Ascii(largeInput + "Hello", new ArraySegment<byte>(requestArray, 0, requestArray.Length));
+
+                await readerTask;
+
+                var count = await stream.ReadAsync(new byte[1], 0, 1);
+                Assert.Equal(0, count);
+            }
         }
     }
 }
