+++ test/Microsoft.AspNetCore.Server.Kestrel.Core.Tests/TestHelpers/AssertExtensions.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+using System.Text;
 using Microsoft.AspNetCore.Server.Kestrel.Internal.System;
 using Xunit.Sdk;
 
@@ -21,6 +23,16 @@ namespace Xunit
                 {
                     throw new XunitException($@"Expected byte at index {i} to be '{expected[i]}' but was '{actual[i]}'");
                 }
+            }
+        }
+
+        public static void Ascii(string expected, ArraySegment<byte> actual)
+        {
+            var bytes = Encoding.ASCII.GetBytes(expected);
+            Assert.Equal(bytes.Length, actual.Count);
+            for (var index = 0; index < bytes.Length; index++)
+            {
+                Assert.Equal(bytes[index], actual.Array[actual.Offset + index]);
             }
         }
     }
